change react-router init to new (router v6) style
