+++ services/web/admin/src/App.tsx
@@ -1,8 +1,7 @@
 import { useAppSelector, useInitAppAdmin } from '@m-cafe-app/frontend-logic/admin/hooks';
-import { AppRoutes } from './AppRoutes';
-import { Loading, AppContent, Wrapper, Header } from 'shared/components';
-import { collapseExpanded } from '@m-cafe-app/frontend-logic/utils';
-import { StaffSidebar } from 'shared/staffComponents';
+import { Loading } from 'shared/components';
+import { RouterProvider } from 'react-router-dom';
+import { router } from './router';
 
 
 export const App = () => {
@@ -11,33 +10,11 @@ export const App = () => {
 
   const uiSettingsHash = useAppSelector(state => state.settings.uiSettingsHash);
   const fixedLocsHash = useAppSelector(state => state.fixedLocs.locsHash);
-  const user = useAppSelector(state => state.user);
 
   if (!uiSettingsHash || !fixedLocsHash)
     return <Loading size='medium'/>;
 
-  if (!user.phonenumber) {
-    return (
-      <>
-        {/* <Notification/> */}
-        <Wrapper id='app-wrapper' onClick={() => collapseExpanded()}>
-          <Header loginNecessary={true}/>
-          <AppContent />
-        </Wrapper>
-      </>
-    );
-  } else {
-    return (
-      <>
-        <Wrapper id='app-wrapper' onClick={() => collapseExpanded()}>
-          <Header loginNecessary={true}/>
-          <AppContent>
-            {/* <Notification/> */}
-            <AppRoutes/>
-          </AppContent>
-          <StaffSidebar/>
-        </Wrapper>
-      </>
-    );
-  }
+  return (
+    <RouterProvider router={router} />
+  );
 };
