+++ services/web/admin/src/index.tsx
@@ -2,8 +2,6 @@ import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import store from '@m-cafe-app/frontend-logic/admin';
-import { BrowserRouter as Router } from 'react-router-dom';
-import { Routes, Route } from 'react-router-dom';
 import { App } from './App';
 
 import 'shared/styles/main.scss';
@@ -20,14 +18,9 @@ if (rootElement) {
     <React.StrictMode>
       <Suspense fallback={<Loading size='medium'/>}>
         <Provider store={store}>
-          <Router>
-            <Routes>
-              <Route path="/*" element={<App />} />
-            </Routes>
-          </Router>
+          <App/>
         </Provider>
       </Suspense>
     </React.StrictMode>,
   );  
-
 }
